add pose priors and first frame floor constraints back into marker solver
Summary:
Added functionality to allow locking the floor contact locators to y=0 in the first frame for better foot calibration.
Added functionality to add pose constraints for first frame in calibration (i.e. to enforce straigh arms/legs/spine in case it is known the sequence starts with a t-pose)
This will lead to better calibration results.

Reviewed By: jeongseok-meta

Differential Revision: D74821953